fix: 优化PicGO调用upyun平台，新增测试
