+++ .gitignore
@@ -1,3 +1,4 @@
 .idea
 node_modules
 dist
+test-dist
