+++ package.json
@@ -1,14 +1,13 @@
 {
   "name": "sy-picgo",
-  "version": "1.1.3",
+  "version": "1.1.5",
   "description": "",
   "scripts": {
     "serve": "pnpm watch",
     "dev": "pnpm watch",
     "build": "bash scripts/build.sh",
     "watch": "node esbuild.mjs --watch",
-    "test": "node dist/webview/index.js",
-    "testExtension": "node dist/extension.js"
+    "test": "esbuild test/test.ts --outfile=test-dist/test.js && node test-dist/test.js"
   },
   "keywords": [],
   "author": "",
