--- src/webview/pages/index.tsx
@@ -1,5 +1,0 @@
-import * as React from "react"
-import * as Server from "react-dom/server"
-
-let Greet = () => <h1>Hello, world!</h1>
-console.log(Server.renderToString(<Greet />))
